+++ src/Effectful/Tracing/Interpreter/NoOp.hs
@@ -30,6 +30,7 @@ import Effectful.Tracing.Effect
       , GetActiveSpan
       , RecordException
       , SetStatus
+      , WithLinkedRoot
       , WithSpan
       )
   )
@@ -44,6 +45,7 @@ import Effectful.Tracing.Effect
 runTracerNoOp :: Eff (Tracer : es) a -> Eff es a
 runTracerNoOp = interpret $ \env -> \case
   WithSpan _ _ action -> localSeqUnlift env (\unlift -> unlift action)
+  WithLinkedRoot _ action -> localSeqUnlift env (\unlift -> unlift action)
   AddAttribute _ _ -> pure ()
   AddAttributes _ -> pure ()
   AddEvent _ _ -> pure ()